requirements.txt
streamlit==1.37.0
pandas==2.2.2
scikit-learn==1.5.1
rapidfuzz==3.9.6
pdfplumber==0.11.4
python-dotenv==1.0.1
google-generativeai==0.7.2